NumberGuessingGame.java
Hi, I am Prateek Kumar. This is for my codesoft internship. In this project I  made Number Guessing Game in JAVA programming . In this game user has to enter a number from 1 to 100 . In these number there is a random number user have to guess . If user guess the correct number then output will be you have guess right number in N times.